requirements.txt
appdirs==1.4.0
Cython==0.25.2
Kivy==1.9.1
Kivy-Garden==0.1.4
Naked==0.1.31
packaging==16.8
pycrypto==2.6.1
pygame==1.9.3
PyKCS11==1.4.1
pyparsing==2.1.10
PyYAML==3.12
requests==2.13.0
shellescape==3.4.1
six==1.10.0
